KomurSiniflandirma/Dayi/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
assemblyai==0.23.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chromedriver-py==124.0.6367.91
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.2
cycler==0.10.0
Cython==3.0.10
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.10.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==2.10
iopath==0.1.10
ipywidgets==8.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
keras==3.3.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.36
lapx==0.5.8
layoutparser==0.3.4
libclang==18.1.1
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.7.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.30.3
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.0
packaging==23.2
pandas==2.2.0
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.11.0
pillow==10.2.0
pillow_heif==0.15.0
playwright==1.44.0
plotly==5.20.0
portalocker==2.8.2
protobuf==4.23.4
psutil==5.9.5
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.2.7
pytz==2024.1
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
roboflow==1.1.28
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.10.1
seaborn==0.13.2
SQLAlchemy==2.0.29
streamlit==1.32.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
timm==0.9.12
tokenizers==0.15.2
toml==0.10.2
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.37.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.11
unstructured==0.12.6
urllib3==1.26.18
watchdog==4.0.0
websockets==10.4
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
